Simplify mounting/unmounting documents
