+++ web/packages/teleterm/src/ui/Document/index.ts
@@ -19,3 +19,4 @@
 import Document from './Document';
 
 export default Document;
+export { ForegroundSession } from './Document';
